Extract FetchMapiUser component
