+++ src/App.jsx
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import CallContract from './components/CallContract'
 import DeployContract from './components/DeployContract'
 import FetchUser from './components/Fetchuser'
+import FetchMapiUser from './components/FetchMapiUser'
 import GetAccount from './components/GetAccount'
 import GetTokenOwner from './components/GetTokenOwner'
 
@@ -18,7 +19,9 @@ function App() {
 
       <p>Step 2: Programmatically fetch user</p>
 
-      <FetchUser address={address} />
+      <FetchUser />
+
+      <FetchMapiUser address={address} />
 
       <p>Step 3: Deploy web3 contract</p>
